requirements.txt
# Dependencies
PyMuPDF>=1.24.0
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
PyYAML>=6.0
httpx>=0.25.0

google-generativeai>=0.8.0

# LanceDB vector store + embeddings (sentence-transformers)
lancedb>=0.5.0
sentence-transformers>=2.2.0

# Microsoft Agent Framework (MAF) for agentic retrieval (optional; use_maf_retrieval in config)
agent-framework>=1.0.0rc4; python_version >= "3.10"

pytest>=7.4.0
pytest-cov>=4.1.0
